experiment: render via sigma.js #147
Problem: filteredGraph 'instance' cannot be sent as is to client side, export / import is necessary.
As a consequence, it is necessary to update the graph instance manually client side (add new nodes,
update attributes of existing ones, delete obsolete nodes) at each update. Not the most efficient.
In the next commits, graph display will be managed with d3, as data joins will help.